Full file system refactor
