+++ src/public/assets/C logo.png
