+++ src/public/css/detailed.css
