+++ src/public/css/generate_teams.css
