+++ src/views/generate_teams.ejs
