+++ src/views/instructor_home.ejs
@@ -12,7 +12,7 @@
 <body>
 
 <header>
-    <img src="/Assets/concordia logo.png" alt="Concordia Logo">
+    <img src="/assets/concordia logo.png" alt="Concordia Logo">
     <h1>Welcome to the Peer Evaluation App</h1>
 </header>
