+++ src/views/manage_teams.ejs
@@ -10,7 +10,7 @@
 <body>
 
 <header>
-    <img src="/Assets/concordia logo.png" alt="Concordia Logo">
+    <img src="/assets/concordia logo.png" alt="Concordia Logo">
     <h1>Manage Teams</h1>
 </header>
