+++ src/views/partials/instructor_navbar.ejs
@@ -1,6 +1,6 @@
 <link rel="stylesheet" href="/css/instructor_navbar.css">
 <nav class="navbar">
-    <img src="/Assets/concordia logo.png" alt="Concordia Logo">
+    <img src="/assets/concordia logo.png" alt="Concordia Logo">
     <ul class="navbar-list">
         <!--Will change these as needed -->
         <li><a href="/instructor_index">Home</a></li>
