+++ src/views/partials/student_navbar.ejs
@@ -1,7 +1,7 @@
 <link rel="stylesheet" href="/css/instructor_navbar.css">
 
 <nav class="navbar">
-    <img src="/Assets/concordia logo.png" alt="Concordia Logo">
+    <img src="/assets/concordia logo.png" alt="Concordia Logo">
     <ul class="navbar-list">
         <!--Will change these as needed -->
         <li><a href="/student_index">Home</a></li>
