+++ src/views/register.ejs
