+++ src/views/student_index.ejs
