+++ src/views/team_management.ejs
@@ -10,7 +10,7 @@
 <body>
 
 <header>
-    <img src="/Assets/concordia logo.png" alt="Concordia Logo">
+    <img src="/assets/concordia logo.png" alt="Concordia Logo">
     <h1>Team management</h1>
 </header>
